# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Back End API",
        "description": "Ecommerce-app endpoint",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/order-registration": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Register a new order",
                "description": "Creates New Order",
                "operationId": "a2580df9ab7f2e7819eae9f0e6dab40b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shipping_address_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "billing_address_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "credit_card"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The Products was ordered successfuly"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders for the authenticated user",
                "description": "Get Orders",
                "operationId": "54561fcdcc8f3f1e8cfd59f2bb99df7c",
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get orders successfuly"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderResource": {
                "title": "OrderResource Schema",
                "properties": {
                    "order_id": {
                        "type": "integer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "code": {
                        "type": "string"
                    },
                    "shippingAddress": {
                        "type": "string"
                    },
                    "itemCount": {
                        "type": "integer"
                    },
                    "totalPrice": {
                        "type": "number"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderedProductsResource"
                        }
                    },
                    "orderStatus": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "done": {
                                    "type": "boolean"
                                },
                                "createdDate": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "OrderedProductsResource": {
                "title": "OrderedProductsResource Schema",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "product_title": {
                        "type": "string"
                    },
                    "product_quantity": {
                        "type": "integer"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "product_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "product_color": {
                        "type": "string"
                    },
                    "product_image": {
                        "type": "string"
                    },
                    "product_size": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource Schema",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}